% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichStep.R
\docType{class}
\name{EnrichStep-class}
\alias{EnrichStep-class}
\title{Base class of this package}
\description{
This class is inherit from \code{Step} in pipeFrame package,
no more method is extended or override. Please see \code{Step}
class for detail.
}
