% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenBackground.R
\docType{methods}
\name{GenBackground}
\alias{GenBackground}
\alias{enrichGenBackground,Step-method}
\alias{enrichGenBackground}
\alias{genBackground}
\title{Generate background regions and reset foreground regions sizes}
\usage{
enrichGenBackground(prevStep, inputForegroundBed = NULL, genome = NULL,
  outputForegroundBed = NULL, outputBackgroundBed = NULL,
  outputRegionBed = NULL, rangeLen = 1000, sampleNumb = 10000, ...)

\S4method{enrichGenBackground}{Step}(prevStep, inputForegroundBed = NULL,
  genome = NULL, outputForegroundBed = NULL,
  outputBackgroundBed = NULL, outputRegionBed = NULL,
  rangeLen = 1000, sampleNumb = 10000, ...)

genBackground(inputForegroundBed, genome = NULL,
  outputForegroundBed = NULL, outputBackgroundBed = NULL,
  outputRegionBed = NULL, rangeLen = 1000, sampleNumb = 10000, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It has to be the return value of upstream process from other package like esATAC}

\item{inputForegroundBed}{\code{Character} scalar.
Foreground BED file directory.}

\item{genome}{\code{Character} scalar.
Bioconductor supported genome like "hg19", "mm10", etc.
Default: NULL (call function like \code{setGenome("hg19")} first after library this package)}

\item{outputForegroundBed}{\code{Character} scalar.
Reshaped foreground region BED file directory.
Default: NULL (generate base on inputForegroundBed)}

\item{outputBackgroundBed}{\code{Character} scalar.
Reshaped background region BED file directory
Default: NULL (generate base on inputForegroundBed)}

\item{outputRegionBed}{\code{Character} scalar.
Merged foreground and background BED file.
Default: NULL (generate base on inputForegroundBed)}

\item{rangeLen}{\code{Character} scalar. The length of forground regions will be set. Default: 1000}

\item{sampleNumb}{\code{numeric} scalar. The number of background regions will be sampled. Default: 10000}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichTF-class}} object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Use uniform distribution to generate background regions on genome.
The foreground regions' sizes will be unified into set length in argument
}
\details{
Use uniform distribution to generate background regions on genome.
The foreground regions' sizes will be unified into set length in argument
}
\examples{

foregroundBedPath <- system.file(package = "enrichTF", "extdata","testForeGround.bed")
gen <- genBackground(inputForegroundBed = foregroundBedPath,genome = "hg19")
}
\seealso{
\code{\link{regionConnectTargetGene}}
\code{\link{findMotifsInRegions}}
\code{\link{tfsEnrichInRegions}}
}
\author{
Zheng Wei
}
