% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneOntology.R
\name{GeneOntology}
\alias{GeneOntology}
\alias{enrichGeneOntology,Step-method}
\alias{enrichGeneOntology}
\alias{geneOntology}
\title{Gene ontology enrichment analysis for provided gene list}
\usage{
enrichGeneOntology(
  prevStep,
  inputTxt = NULL,
  outputTxt = NULL,
  outputPdf = NULL,
  orgDb = NULL,
  keyType = "SYMBOL",
  ...
)

\S4method{enrichGeneOntology}{Step}(
  prevStep,
  inputTxt = NULL,
  outputTxt = NULL,
  outputPdf = NULL,
  orgDb = NULL,
  keyType = "SYMBOL",
  ...
)

geneOntology(
  inputTxt,
  outputTxt = NULL,
  outputPdf = NULL,
  orgDb = NULL,
  keyType = "SYMBOL",
  ...
)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
This parameter is available when the upstream step function
(printMap() to see the previous functions)
have been sucessfully called.
Accepted value can be the object return by any step function or be feed by
\code{\%>\%} from last step function.}

\item{inputTxt}{\code{Character} scalar.
Gene list text file. All gene names are in one column.}

\item{outputTxt}{\code{Character} scalar.
Gene ontology enrichment analysis result table.
Each row contain one gene ontology information.}

\item{outputPdf}{\code{Character} scalar.
Gene ontology enrichment analysis result figure.
It contains gene ontology network.}

\item{orgDb}{\code{Character} scalar.
Bioconductor OrgDb object name for gene ontology enrichment analysis.}

\item{keyType}{\code{Character} scalar.
Gene name type include "SYMBOL" and "ENSEMBLE"}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}} object
(\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
User provide target gene list of forground region.
This function will call function for gene ontology enrichment analysis
}
\details{
Currently, this function call enrichGO from package clusterProfiler to implement this funtion.
}
\examples{

genelist.txt <- system.file(package = "enrichTF", "extdata","genelist.txt")
geneOntology(inputTxt = genelist.txt, orgDb = "org.Hs.eg.db")

}
\seealso{
\code{\link{regionConnectTargetGene}}
\code{\link{enrichRegionConnectTargetGene}}
}
\author{
Zheng Wei
}
