% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMotifsInRegions.R
\docType{methods}
\name{MotifsInRegions}
\alias{MotifsInRegions}
\alias{enrichFindMotifsInRegions,Step-method}
\alias{enrichMotifsInRegions}
\alias{findMotifsInRegions}
\alias{motifsInRegions}
\title{Find motifs in all provided regions}
\usage{
enrichFindMotifsInRegions(prevStep, inputRegionBed = NULL,
  outputRegionMotifBed = NULL, motifRc = c("integrate", "jaspar",
  "pwmfile"), inputPwmFile = NULL, genome = NULL, ...)

\S4method{enrichFindMotifsInRegions}{Step}(prevStep,
  inputRegionBed = NULL, outputRegionMotifBed = NULL,
  motifRc = c("integrate", "jaspar", "pwmfile"), inputPwmFile = NULL,
  genome = NULL, ...)

findMotifsInRegions(inputRegionBed, outputRegionMotifBed,
  motifRc = c("integrate", "jaspar", "pwmfile"), inputPwmFile = NULL,
  genome = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It has to be the return value of upstream process from \code{\link{genBackground}} and \code{\link{enrichGenBackground}}}

\item{inputRegionBed}{\code{Character} scalar.
Regions BED file including foreground and background}

\item{outputRegionMotifBed}{\code{Character} scalar.
Regions BED file with motif candidates.
Default: NULL (generate base on inputForegroundBed)}

\item{motifRc}{\code{Character} scalar.
Motif Resources can be one of "integrate"
(integrated by us and can be download from internet automatically
if call function like \code{setGenome("hg19")} first after library this package),
"jaspar" package JASPAR2018,
and "pwmfile" (user defined PWM file. when this is set, inputPwmFile is required)}

\item{inputPwmFile}{\code{Character} scalar.
when "pwmfile" is set for motifRc, users use this argument to provid PWM file directory}

\item{genome}{\code{Character} scalar.
Bioconductor supported genome like "hg19", "mm10", etc.
Default: NULL (call function like \code{setGenome("hg19")} first after library this package)}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichTF-class}} object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Use prepared motifs pwm to scan regions and give motif candidates in regions.
}
\details{
Connect foreground and background regions to targetGene
}
\examples{
setGenome("hg19")
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testForeGround.bed")
gen <- genBackground(inputForegroundBed = foregroundBedPath)
findMotif <- enrichFindMotifsInRegions(gen,motifRc="integrate")
}
\seealso{
\code{\link{genBackground}}
\code{\link{findMotifsInRegions}}
\code{\link{tfsEnrichInRegions}}
}
\author{
Zheng Wei
}
