% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method.R
\name{PECA_TF_enrich}
\alias{PECA_TF_enrich}
\title{TF enrich with PECA2 mothod}
\usage{
PECA_TF_enrich(inputForegroundBed, genome, threads = 2, ...)
}
\arguments{
\item{inputForegroundBed}{\code{Character} scalar.
Foreground BED file directory.}

\item{genome}{\code{Character} scalar.
Bioconductor supported genome like "hg19", "mm10", etc.}

\item{threads}{\code{Numeric} scalar.
The max number of threads that can be used by each step of the pipeline}

\item{...}{Additional arguments to set arguments for each Steps.
See below for details.}
}
\value{
An invisible \code{list} contain all four steps \code{EnrichTF} objects
}
\description{
TF enrich with PECA2 mothod
}
\details{
There four steps in this pipeline: GenBackground, RegionConnectTarget,
FindMotifsInRegions and TFsEnrichInRegions. For instance, if you want to change
the number of background regions (\code{sampleNumb}) into 1000,
you can add argument \code{GenBackground.sampleNumb = 1000} to the function in this way:
\code{PECA_TF_enrich(inputForegroundBed = "your/file.bed",genome="hg19",GenBackground.sampleNumb = 1000)}.
The number of arguments is not limited so you can add other arguments with \code{StepName.argumentName} in the same way.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed")
PECA_TF_enrich(inputForegroundBed = foregroundBedPath, genome = "testgenome")
}
\references{
Duren Z.,.....PNAS
}
\author{
Zheng Wei
}
