% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSampleReport.R
\docType{methods}
\name{SingleSampleReport}
\alias{SingleSampleReport}
\alias{enrichSingleSampleReport,Step-method}
\alias{enrichSingleSampleReport}
\title{Tissue's open conservation of the given region}
\usage{
enrichSingleSampleReport(prevStep, htmlOutput = NULL, ...)

\S4method{enrichSingleSampleReport}{Step}(prevStep, htmlOutput = NULL,
  ...)
}
\arguments{
\item{htmlOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on bedInput)}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
User provide region through a BED file.
This function will provide tissue's open conservation analysis for these region.
}
\details{
We collected 201 DNase-seq or ATAC-seq sample from ENCODE and calculate their open level value.
They can be download and install automatically. So users do not need to configure themselves.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed.gz")
rs <- unzipAndMergeBed(bedInput = foregroundBedPath) \%>\%
         enrichTissueOpennessSpecificity
}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
