% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueOpennessConserve.R
\docType{methods}
\name{TissueOpennessConserve}
\alias{TissueOpennessConserve}
\alias{enrichTissueOpennessConserve,Step-method}
\alias{enrichTissueOpennessConserve}
\alias{tissueOpennessConserve}
\title{Conservation score of openness across tissues for the given regions}
\usage{
enrichTissueOpennessConserve(prevStep, bedInput = NULL,
  openConserveBedInput = NULL, bedOutput = NULL,
  distrPdfOutput = NULL, ...)

\S4method{enrichTissueOpennessConserve}{Step}(prevStep, bedInput = NULL,
  openConserveBedInput = NULL, bedOutput = NULL,
  distrPdfOutput = NULL, ...)

tissueOpennessConserve(bedInput, openConserveBedInput = NULL,
  bedOutput = NULL, distrPdfOutput = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
This parameter is available when the upstream step function
(printMap() to check the previous functions)
has been sucessfully called.
Accepted value can be the object returned by any step function or be fed by
\code{\%>\%} from last step function.}

\item{bedInput}{\code{Character} scalar.
The directory of input BED file for analysis.}

\item{openConserveBedInput}{\code{Character} scalar.
The openness conservation score file for analysis. The first three columns are chromosome name, start and end positions,
and the remaining two columns are the region name and conservation score across tissues.}

\item{bedOutput}{\code{Character} scalar.
The output file directory of merged BED files.
Default: NULL (generated based on bedInput)}

\item{distrPdfOutput}{\code{Character} scalar.
The openness conservation distribution figure will be provided in a PDF file.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (based on \code{\link{Step-class}}) scalar for downstream analysis.
}
\description{
Users provide region positions through a BED file.
This function will provide the conservation score of openness across tissues for these given regions.
}
\details{
We collected 201 DNase-seq/ATAC-seq samples from ENCODE and calculated their openness levels.
These can be downloaded and installed automatically. So users do not need to configure by themselves.
}
\examples{

foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed")
tissueOpennessConserve(bedInput = foregroundBedPath)
}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
