% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueOpennessSpecificity.R
\docType{methods}
\name{TissueOpennessSpecificity}
\alias{TissueOpennessSpecificity}
\alias{enrichTissueOpennessSpecificity,Step-method}
\alias{enrichTissueOpennessSpecificity}
\alias{tissueOpennessSpecificity}
\title{Tissue-specific openness scores for the given regions}
\usage{
enrichTissueOpennessSpecificity(prevStep, bedInput = NULL,
  openBedInput = NULL, sampleTxtInput = NULL, bedOutput = NULL,
  distPdfOutput = NULL, heatmapPdfOutput = NULL,
  sampleTxtOutput = NULL, ...)

\S4method{enrichTissueOpennessSpecificity}{Step}(prevStep,
  bedInput = NULL, openBedInput = NULL, sampleTxtInput = NULL,
  bedOutput = NULL, distPdfOutput = NULL, heatmapPdfOutput = NULL,
  sampleTxtOutput = NULL, ...)

tissueOpennessSpecificity(bedInput, openBedInput = NULL,
  sampleTxtInput = NULL, bedOutput = NULL, distPdfOutput = NULL,
  heatmapPdfOutput = NULL, sampleTxtOutput = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
This parameter is available when the upstream step functions
(printMap() to check the previous functions)
have been sucessfully called.
Accepted value can be the object returned by any step function or be fed by
\code{\%>\%} from last step function.}

\item{bedInput}{\code{Character} scalar.
The directory of input BED file for analysis.}

\item{openBedInput}{\code{Character} scalar.
The openness level file of different tissues for analysis. The first three columns are chromosome name, start and end positions,
The remaining columns are the openness values for each tissue.
The order of tissues should be consistent with the file provided by sampleTxtInput.}

\item{sampleTxtInput}{\code{Character} scalar.
The tissue annotation of samples in the file provided by openBedInput.
There are 4 columns seperated by tab. The first column indicates the order.
The second column is the tissue annotation in detail.
The third column is the tissue name.
The forth column is the accession number in its original database such as ENCODE project.}

\item{bedOutput}{\code{Character} scalar.
The output file directory of merged BED files.
Default: NULL (generated based on bedInput)}

\item{distPdfOutput}{\code{Character} scalar.
The openness level distribution figure for each tissue will be provided in PDF file.
It is shown in descending order.}

\item{heatmapPdfOutput}{\code{Character} scalar.
The openness level hierarchical clustering heatmap base on regions and tissues will be provided in this PDF file.
The corresponding heatmap data will be stored in the Rdata format at the same directory.}

\item{sampleTxtOutput}{\code{Character} scalar.
In this file, there are five columns seperated by tab.
The first four columns are the same with sampleTxtInput:
The first column is the order number.
The second column is the tissue detail information.
The third column is the tissue name.
The forth column is the code from source project such as ENCODE project.
The last column is the median value of the openness level for each tissue.
The table is in descending order of last column}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Users provide region positions through a BED file.
This function will provide tissue-specific openness degree of the given regions.
Median value of Openness level,  openness distribution and clustering result (heatmap)
based on tissues and regions will be caculated.
}
\details{
We collect 201 DNase-seq or ATAC-seq samples from ENCODE and calculate their openness level value.
They can be downloaded and installed automatically. So users do not need to configure by themselves.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed")
tissueOpennessSpecificity(bedInput = foregroundBedPath)

}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
