% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnzipAndMergeBed.R,
%   R/TissueOpennessSpecificity.R
\docType{methods}
\name{UnzipAndMergeBed}
\alias{UnzipAndMergeBed}
\alias{enrichUnzipAndMergeBed,Step-method}
\alias{enrichUnzipAndMergeBed}
\alias{unzipAndMergeBed}
\title{Generate background regions and reset the size of
foreground regions}
\usage{
enrichUnzipAndMergeBed(prevStep, bedInput = NULL, bedOutput = NULL,
  ...)

\S4method{enrichUnzipAndMergeBed}{Step}(prevStep, bedInput = NULL,
  bedOutput = NULL, ...)

unzipAndMergeBed(bedInput, bedOutput = NULL, ...)

enrichUnzipAndMergeBed(prevStep, bedInput = NULL, bedOutput = NULL,
  ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It needs to be the return value of upstream process
from other packages, such as esATAC.}

\item{bedInput}{\code{Character} scalar or vector.
The directory of region BED files for analysis.
BED, BED.gz, BED.bz2 formats are supported.}

\item{bedOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on bedInput)}

\item{...}{Additional arguments, currently unused.}

\item{prevStep}{\code{\link{Step-class}} object scalar.
It needs to be the return value of upstream process
from other packages, such as esATAC.}

\item{bedInput}{\code{Character} scalar or vector.
The directory of region BED files for analysis.
BED, BED.gz, BED.bz2 formats are supported.}

\item{bedOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on bedInput)}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.

An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Use uniform distribution to generate background
sequence regions from genome.
The size of foreground regions will be unified into the
length specified in argument.

Use uniform distribution to generate background
sequence regions from genome.
The size of foreground regions will be unified into the
length specified in argument.
}
\details{
All compressed files will be de-compressed.
Only first 3 columns (chromasomes, start and end) will be collected.
All BED files will be merged into one BED file.

All compressed files will be de-compressed.
Only first 3 columns (chromasomes, start and end) will be collected.
All BED files will be merged into one BED file.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed.gz")
gen <- unzipAndMergeBed(bedInput = foregroundBedPath)
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed.gz")
gen <- unzipAndMergeBed(bedInput = foregroundBedPath)
}
\seealso{
\code{\link{genBackground}}

\code{\link{genBackground}}
}
\author{
Zheng Wei

Zheng Wei
}
