% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenBackground.R
\name{GenBackground}
\alias{GenBackground}
\alias{enrichGenBackground,Step-method}
\alias{enrichGenBackground}
\alias{genBackground}
\title{Generate background regions and reset the size of
foreground regions}
\usage{
enrichGenBackground(
  prevStep,
  inputForegroundBed = NULL,
  inputBackgroundBed = NULL,
  genome = NULL,
  outputForegroundBed = NULL,
  outputBackgroundBed = NULL,
  outputRegionBed = NULL,
  regionLen = 1000,
  sampleNumb = 10000,
  ...
)

\S4method{enrichGenBackground}{Step}(
  prevStep,
  inputForegroundBed = NULL,
  inputBackgroundBed = NULL,
  genome = NULL,
  outputForegroundBed = NULL,
  outputBackgroundBed = NULL,
  outputRegionBed = NULL,
  regionLen = 1000,
  sampleNumb = NULL,
  ...
)

genBackground(
  inputForegroundBed,
  inputBackgroundBed = NULL,
  genome = NULL,
  outputForegroundBed = NULL,
  outputBackgroundBed = NULL,
  outputRegionBed = NULL,
  regionLen = 1000,
  sampleNumb = NULL,
  ...
)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It needs to be the return value of upstream process
from other packages, such as esATAC.}

\item{inputForegroundBed}{\code{Character} scalar.
The directory of foreground BED file.}

\item{inputBackgroundBed}{\code{Character} scalar.
The directory of backgroud BED file.
Default: NULL to generate random region from genome.}

\item{genome}{\code{Character} scalar.
Bioconductor supported genome such as "hg19", "mm10", etc.
Default: NULL (e.g. after library (enrichTF),
you can call function \code{setGenome("hg19")})}

\item{outputForegroundBed}{\code{Character} scalar.
The BED file directory of reshaped foreground regions.
Default: NULL (generated base on inputForegroundBed)}

\item{outputBackgroundBed}{\code{Character} scalar.
The BED file directory of reshaped background regions.
Default: NULL (generated base on inputForegroundBed)}

\item{outputRegionBed}{\code{Character} scalar.
Foreground and background merged BED files.
Default: NULL (generated base on inputForegroundBed)}

\item{regionLen}{\code{Character} scalar.
It sets the length of forground sequence regions.
Default: 1000}

\item{sampleNumb}{\code{numeric} scalar.
It sets the number of background regions that
will be sampled. Default: 10000}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Use uniform distribution to generate background
sequence regions from genome.
The size of foreground regions will be unified into the
length specified in argument.
}
\details{
Use uniform distribution to generate background sequence
regions from genome.
The size of foreground regions will be unified into the
 length specified in argument.
}
\examples{
setGenome("testgenome") #Use "hg19","hg38",etc. for your application
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed")
gen <- genBackground(inputForegroundBed = foregroundBedPath)
}
\seealso{
\code{\link{regionConnectTargetGene}}
\code{\link{findMotifsInRegions}}
\code{\link{tfsEnrichInRegions}}
}
\author{
Zheng Wei
}
