% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneOntology.R
\docType{methods}
\name{GeneOntology}
\alias{GeneOntology}
\alias{enrichGeneOntology,Step-method}
\alias{enrichGeneOntology}
\alias{geneOntology}
\title{Connect regions with their target genes}
\usage{
enrichGeneOntology(prevStep, inputTxt = NULL, outputTxt = NULL,
  outputPdf = NULL, orgDb = NULL, keyType = "SYMBOL", ...)

\S4method{enrichGeneOntology}{Step}(prevStep, inputTxt = NULL,
  outputTxt = NULL, outputPdf = NULL, orgDb = NULL,
  keyType = "SYMBOL", ...)

geneOntology(inputTxt, outputTxt = NULL, outputPdf = NULL,
  orgDb = NULL, keyType = "SYMBOL", ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It needs to be the return value of upstream process from
\code{\link{genBackground}} or \code{\link{enrichGenBackground}}
when it is not used in a pipeline.  If it is used in a pipeline or
\code{\%>\%} is applied on this function, any steps in this package
is acceptable.}

\item{inputTxt}{\code{Character} scalar.
The BED file directory of foreground regions.}

\item{outputTxt}{\code{Character} scalar.
The BED file directory of background regions.}

\item{outputPdf}{\code{Character} scalar.
The BED file directory of target genes connecting with foreground regions,
which are derived from PECA model.
Default: NULL (generated base on inputForegroundBed)}

\item{orgDb}{\code{Character} scalar.
The BED file directory of target genes connecting with background regions,
which are derived from PECA model.
Default: NULL (generated base on inputBackgroundBed)}

\item{keyType}{\code{Character} scalar.
The BED file directory of target genes which are predicted from PECA.
Default: NULL (e.g. after \code{library (enrichTF)}, you can call
function \code{setGenome("hg19")})}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}} object
(\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Connect foreground and background regions to their target genes,
which is predicted from PECA model.
}
\details{
Connect foreground and background regions to target genes,
which are predicted from PECA.
}
\examples{

genelist.txt <- system.file(package = "enrichTF", "extdata","genelist.txt")
geneOntology(inputTxt = genelist.txt, orgDb = "org.Hs.eg.db")

}
\seealso{
\code{\link{genBackground}}
\code{\link{findMotifsInRegions}}
\code{\link{tfsEnrichInRegions}}
}
\author{
Zheng Wei
}
