% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectTargetGene.R
\docType{methods}
\name{RegionConnectTargetGene}
\alias{RegionConnectTargetGene}
\alias{enrichRegionConnectTargetGene,Step-method}
\alias{enrichRegionConnectTargetGene}
\alias{regionConnectTargetGene}
\title{Connect Region and Target Gene}
\usage{
enrichRegionConnectTargetGene(prevStep, inputForegroundBed = NULL,
  inputBackgroundBed = NULL, outputForegroundBed = NULL,
  outputBackgroundBed = NULL, regularGeneCorrBed = NULL,
  enhancerRegularGeneCorrBed = NULL, ...)

\S4method{enrichRegionConnectTargetGene}{Step}(prevStep,
  inputForegroundBed = NULL, inputBackgroundBed = NULL,
  outputForegroundBed = NULL, outputBackgroundBed = NULL,
  regularGeneCorrBed = NULL, enhancerRegularGeneCorrBed = NULL, ...)

regionConnectTargetGene(inputForegroundBed, inputBackgroundBed,
  outputForegroundBed = NULL, outputBackgroundBed = NULL,
  regularGeneCorrBed = NULL, enhancerRegularGeneCorrBed = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It has to be the return value of upstream process from \code{\link{genBackground}} and \code{\link{enrichGenBackground}}}

\item{inputForegroundBed}{\code{Character} scalar.
Foreground BED file directory.}

\item{inputBackgroundBed}{\code{Character} scalar.
Background BED file directory.}

\item{outputForegroundBed}{\code{Character} scalar.
Gene connected foreground region BED file directory.
Default: NULL (generate base on inputForegroundBed)}

\item{outputBackgroundBed}{\code{Character} scalar.
Gene connected background region BED file directory
Default: NULL (generate base on inputForegroundBed)}

\item{enhancerRegularGeneCorrBed}{\code{Character} scalar.
Enhacer Gene region BED file directory
Default: NULL ((call function like \code{setGenome("hg19")} first after library this package))}

\item{...}{Additional arguments, currently unused.}

\item{regularGeneCorrBed\code{Character}}{scalar.
Gene region BED file directory
Default: NULL ((call function like \code{setGenome("hg19")} first after library this package))}
}
\value{
An invisible \code{\link{EnrichTF-class}} object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
Connect foreground and background regions to targetGene
}
\details{
Connect foreground and background regions to targetGene
}
\examples{
setGenome("hg19")
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testForeGround.bed")
gen <- genBackground(inputForegroundBed = foregroundBedPath)
enrichFindMotifsInRegions(gen,motifRc="integrate")
}
\seealso{
\code{\link{genBackground}}
\code{\link{findMotifsInRegions}}
\code{\link{tfsEnrichInRegions}}
}
\author{
Zheng Wei
}
