% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleSampleReport.R
\name{SingleSampleReport}
\alias{SingleSampleReport}
\alias{enrichSingleSampleReport,Step-method}
\alias{enrichSingleSampleReport}
\title{Final report for single group of regions}
\usage{
enrichSingleSampleReport(prevStep, htmlOutput = NULL, ...)

\S4method{enrichSingleSampleReport}{Step}(prevStep, htmlOutput = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
Any steps object in this package is acceptable when the pipeline is ready.}

\item{htmlOutput}{\code{Character} scalar.
HTML report file directory
Default: NULL (generated base on bedInput)}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
When user call all steps in the pipeline, the final report can be generated.
}
\details{
The report is HTML format. All link in HTML file is the relative directory
in report step folder and other step folder
If user want to move HTML file and keep all link access available,
they should move the whole pipeline folder at the same time.
}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
