% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueOpennessConserve.R
\docType{methods}
\name{TissueOpennessConserve}
\alias{TissueOpennessConserve}
\alias{enrichTissueOpennessConserve,Step-method}
\alias{enrichTissueOpennessConserve}
\alias{tissueOpennessConserve}
\title{Tissue's open conservation of the given region}
\usage{
enrichTissueOpennessConserve(prevStep, bedInput = NULL,
  openConserveBedInput = NULL, bedOutput = NULL,
  distrPdfOutput = NULL, ...)

\S4method{enrichTissueOpennessConserve}{Step}(prevStep, bedInput = NULL,
  openConserveBedInput = NULL, bedOutput = NULL,
  distrPdfOutput = NULL, ...)

tissueOpennessConserve(bedInput, openConserveBedInput = NULL,
  bedOutput = NULL, distrPdfOutput = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
This parameter is available when the upstream step function
(printMap() to see the previous functions)
have been sucessfully called.
Accepted value can be the object return by any step function or be feed by
\code{\%>\%} from last step function.}

\item{bedInput}{\code{Character} scalar.
The directory of region BED file for analysis.}

\item{openConserveBedInput}{\code{Character} scalar.
The open level BED file for analysis. The first three columns are chromosome, start and end,
The remaining columns are region name and conservation score.}

\item{bedOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on bedInput)}

\item{distrPdfOutput}{\code{Character} scalar.
The open conservation distribution figure for each tissue will be provided in PDF file.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
User provide region through a BED file.
This function will provide tissue's open conservation analysis for these region.
}
\details{
We collected 201 DNase-seq or ATAC-seq sample from ENCODE and calculate their open level value.
They can be download and install automatically. So users do not need to configure themselves.
}
\examples{

foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed")
tissueOpennessConserve(bedInput = foregroundBedPath)
}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
