% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueOpennessConserve.R
\name{TissueOpennessConserve}
\alias{TissueOpennessConserve}
\alias{tissueOpennessConserve}
\title{Tissue's open conservation of the given region}
\usage{
enrichTissueOpennessConserve(prevStep, bedInput = NULL,
  openConserveBedInput = NULL, bedOutput = NULL,
  distrPdfOutput = NULL, ...)

tissueOpennessConserve(bedInput, openConserveBedInput = NULL,
  bedOutput = NULL, distrPdfOutput = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It needs to be the return value of upstream process
from other packages, such as esATAC.}

\item{bedInput}{\code{Character} scalar.
The directory of region BED file for analysis.}

\item{openConserveBedInput}{\code{Character} scalar.
The open level BED file for analysis. The first three columns are chromosome, start and end,
The remaining columns are region name and conservation score.}

\item{bedOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on bedInput)}

\item{distrPdfOutput}{\code{Character} scalar.
The open conservation distribution figure for each tissue will be provided in PDF file.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
User provide region through a BED file.
This function will provide tissue's open conservation analysis for these region.
}
\details{
We collected 201 DNase-seq or ATAC-seq sample from ENCODE and calculate their open level value.
They can be download and install automatically. So users do not need to configure themselves.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed.gz")
rs <- unzipAndMergeBed(bedInput = foregroundBedPath) \%>\%
         enrichTissueOpennessSpecificity
}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
