% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TissueOpennessSpecificity.R
\name{TissueOpennessSpecificity}
\alias{TissueOpennessSpecificity}
\alias{enrichTissueOpennessSpecificity,Step-method}
\alias{enrichTissueOpennessSpecificity}
\alias{tissueOpennessSpecificity}
\title{Tissue's open specificity of the given region}
\usage{
enrichTissueOpennessSpecificity(
  prevStep,
  bedInput = NULL,
  openBedInput = NULL,
  sampleTxtInput = NULL,
  bedOutput = NULL,
  distPdfOutput = NULL,
  heatmapPdfOutput = NULL,
  sampleTxtOutput = NULL,
  ...
)

\S4method{enrichTissueOpennessSpecificity}{Step}(
  prevStep,
  bedInput = NULL,
  openBedInput = NULL,
  sampleTxtInput = NULL,
  bedOutput = NULL,
  distPdfOutput = NULL,
  heatmapPdfOutput = NULL,
  sampleTxtOutput = NULL,
  ...
)

tissueOpennessSpecificity(
  bedInput,
  openBedInput = NULL,
  sampleTxtInput = NULL,
  bedOutput = NULL,
  distPdfOutput = NULL,
  heatmapPdfOutput = NULL,
  sampleTxtOutput = NULL,
  ...
)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
This parameter is available when the upstream step function
(printMap() to see the previous functions)
have been sucessfully called.
Accepted value can be the object return by any step function or be feed by
\code{\%>\%} from last step function.}

\item{bedInput}{\code{Character} scalar.
The directory of region BED file for analysis.}

\item{openBedInput}{\code{Character} scalar.
The open level BED file for analysis. The first three columns are chromosome, start and end,
The remaining columns are the open level for each tissue.
The order of tissue should be consistent with the order in the file provided by sampleTxtInput.}

\item{sampleTxtInput}{\code{Character} scalar.
The tissue sample information of in the file provided by openBedInput.
There are 4 columns seperated by tab. The first column is the order number.
The second column is the tissue detail information.
The third column is the tissue name.
The forth column is the code from source project like ENCODE}

\item{bedOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on bedInput)}

\item{distPdfOutput}{\code{Character} scalar.
The open level distribution figure for each tissue will be provided in PDF file.
The order is strong to weak.}

\item{heatmapPdfOutput}{\code{Character} scalar.
The open level hiachical clustering heatmap base on region and tissue will be provided in this PDF file.
The corresponding heatmap data will store at the same directory with suffix .Rdata}

\item{sampleTxtOutput}{\code{Character} scalar.
In this file, there are five columns seperated with tab.
Fist four columns are the same with sampleTxtInput:
The first column is the order number.
The second column is the tissue detail information.
The third column is the tissue name.
The forth column is the code from source project like ENCODE
The last column is the open level median level for each tissue.
The table is in decreasing order of last column}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (\code{\link{Step-class}} based) scalar for downstream analysis.
}
\description{
User provide region through a BED file.
This function will provide tissue's open specificity analysis for this region.
Open level median,  distribution and clustering result (heatmap)
based on tissue and region will be provided.
}
\details{
We collect 201 DNase-seq or ATAC-seq sample from ENCODE and calculate their open level value.
They can be download and install automatically. So users do not need to configure themselves.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed")
tissueOpennessSpecificity(bedInput = foregroundBedPath)

}
\seealso{
\code{\link{unzipAndMergeBed}}
}
\author{
Zheng Wei
}
