% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnzipAndMergeBed.R
\name{UnzipAndMergeBed}
\alias{UnzipAndMergeBed}
\alias{enrichUnzipAndMergeBed,Step-method}
\alias{enrichUnzipAndMergeBed}
\alias{unzipAndMergeBed}
\title{Unzip all zipped BED files and
merge them into one BED file}
\usage{
enrichUnzipAndMergeBed(prevStep, bedInput = NULL, bedOutput = NULL, ...)

\S4method{enrichUnzipAndMergeBed}{Step}(prevStep, bedInput = NULL, bedOutput = NULL, ...)

unzipAndMergeBed(bedInput, bedOutput = NULL, ...)
}
\arguments{
\item{prevStep}{\code{\link{Step-class}} object scalar.
It needs to be the return value of upstream process
from other packages, such as ATAC-seq
peak calling result from esATAC.}

\item{bedInput}{\code{Character} scalar or vector.
The directory of region BED files for analysis.
BED, BED.gz, BED.bz2 formats are supported.}

\item{bedOutput}{\code{Character} scalar.
The BED output file directory of merged BED files.
Default: NULL (generated base on first BED file in  bedInput)}

\item{...}{Additional arguments, currently unused.}
}
\value{
An invisible \code{\link{EnrichStep-class}}
object (inherit from \code{\link{Step-class}}) scalar for downstream analysis.
}
\description{
This function process region BED files in three step:
First, unzip the gzip and bzip2  BED input files.
Second, select first 3 columns of the BED files.
Third, merge the BED files into one BED.
}
\details{
All compressed files will be de-compressed.
Only first 3 columns (chromasomes, start and end) will be collected.
All BED files will be merged into one BED file.
}
\examples{
foregroundBedPath <- system.file(package = "enrichTF", "extdata","testregion.bed.gz")
gen <- unzipAndMergeBed(bedInput = foregroundBedPath)
}
\seealso{
\code{\link{genBackground}}
}
\author{
Zheng Wei
}
